backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
requests==2.32.5
pydantic[email]==2.10.6
python-dotenv==1.0.1
